assignment 3/requirements.txt
cohere==4.32.10
google-generativeai==0.3.0
groq==0.4.1
huggingface-hub==0.19.4
requests==2.31.0
python-dotenv==1.0.0
